// ui/src/api/generated/models/plugin-template-record-list.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { PluginTemplateRecord } from './plugin-template-record'

/**
 *
 * @export
 * @interface PluginTemplateRecordList
 */
export interface PluginTemplateRecordList {
  /**
   * Indicates whether current page is the first page.
   * @type {boolean}
   * @memberof PluginTemplateRecordList
   */
  first: boolean
  /**
   * Indicates whether current page has previous page.
   * @type {boolean}
   * @memberof PluginTemplateRecordList
   */
  hasNext: boolean
  /**
   * Indicates whether current page has previous page.
   * @type {boolean}
   * @memberof PluginTemplateRecordList
   */
  hasPrevious: boolean
  /**
   * A chunk of items.
   * @type {Array<PluginTemplateRecord>}
   * @memberof PluginTemplateRecordList
   */
  items: Array<PluginTemplateRecord>
  /**
   * Indicates whether current page is the last page.
   * @type {boolean}
   * @memberof PluginTemplateRecordList
   */
  last: boolean
  /**
   * Page number, starts from 1. If not set or equal to 0, it means no pagination.
   * @type {number}
   * @memberof PluginTemplateRecordList
   */
  page: number
  /**
   * Size of each page. If not set or equal to 0, it means no pagination.
   * @type {number}
   * @memberof PluginTemplateRecordList
   */
  size: number
  /**
   * Total elements.
   * @type {number}
   * @memberof PluginTemplateRecordList
   */
  total: number
  /**
   * Indicates total pages.
   * @type {number}
   * @memberof PluginTemplateRecordList
   */
  totalPages: number
}